+++ src/lib.rs
@@ -13,6 +13,7 @@ pub mod status;
 pub mod tables;
 pub mod time;
 pub mod uri;
+pub mod versions;
 
 pub use pool::Pool;
 pub use status::Status;
@@ -37,42 +38,5 @@ pub use generated::apr_uint64_t;
 fn initialize_apr() {
     unsafe {
         assert!(generated::apr_initialize() == generated::APR_SUCCESS as i32);
-    }
-}
-
-pub const APU_MAJOR_VERSION: u32 = crate::generated::APU_MAJOR_VERSION;
-pub const APU_MINOR_VERSION: u32 = crate::generated::APU_MINOR_VERSION;
-pub const APU_PATCH_VERSION: u32 = crate::generated::APU_PATCH_VERSION;
-
-pub fn apu_version_string() -> &'static str {
-    unsafe {
-        let ptr = crate::generated::apu_version_string();
-        std::ffi::CStr::from_ptr(ptr).to_str().unwrap()
-    }
-}
-
-pub const APR_MAJOR_VERSION: u32 = crate::generated::APR_MAJOR_VERSION;
-pub const APR_MINOR_VERSION: u32 = crate::generated::APR_MINOR_VERSION;
-pub const APR_PATCH_VERSION: u32 = crate::generated::APR_PATCH_VERSION;
-
-pub fn apr_version_string() -> &'static str {
-    unsafe {
-        let ptr = crate::generated::apr_version_string();
-        std::ffi::CStr::from_ptr(ptr).to_str().unwrap()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_apu_version_string() {
-        apu_version_string();
-    }
-
-    #[test]
-    fn test_apr_version_string() {
-        apr_version_string();
     }
 }
